👷 Add Maven Wrapper
The Maven Wrapper uses plain Java to initiate the Maven build. Every build will
use the same version number as configured in the `maven-wrapper.properties` file.
The specified version does not need to be installed. It will be downloaded
automatically and cached in the `$HOME/.m2/wrapper/dists` directory. The build
will be independent of the Maven version installed by the developer or build
system. Checking out an old state of the repo will use the exact Maven version
from that time. As a result, the build is always reproducable.

Builds on my machine > builds on your machine > builds on CI...

The Maven Wrapper works with Linux, macOS, Windows, Solaris, etc. A local Maven
installation can still be used. The project can still be build within the IDE.
Most IDEs recognize the Maven Wrapper and make use of it.